The HLT branches are not uniform across NanoAOD: https://github.com/cms-nanoAOD/cmssw/issues/528
Some branches may appear in some files but not in others.

For 2016 data was found that the HLT_TkMu50 was missing in some ROOT files from 2016B era. However, when using
HLT_TkMu50_OR_HLT_Mu50 even when the former is missing the later still need to be checked (i.e the file can't be
dismissed as useless).

The solution was then to read the content of HLT_Mu50 in place of HLT_TkMu50 when it is missing, as the OR operator
will remain unchanged.